internal/cgroups: Pull v2 logic into separate files
