+++ internal/cgroups/cgroups.go
@@ -23,16 +23,6 @@
 
 package cgroups
 
-import (
-	"bufio"
-	"fmt"
-	"io"
-	"os"
-	"path"
-	"strconv"
-	"strings"
-)
-
 const (
 	// _cgroupFSType is the Linux CGroup file system type used in
 	// `/proc/$PID/mountinfo`.
@@ -52,27 +42,11 @@ const (
 	// _cgroupCPUCFSPeriodUsParam is the file name for the CGroup CFS period
 	// parameter.
 	_cgroupCPUCFSPeriodUsParam = "cpu.cfs_period_us"
-
-	// _cgroupv2CPUMax is the file name for the CGroup-V2 CPU max and period
-	// parameter.
-	_cgroupv2CPUMax = "cpu.max"
-	// _cgroupFSType is the Linux CGroup-V2 file system type used in
-	// `/proc/$PID/mountinfo`.
-	_cgroupv2FSType = "cgroup2"
 )
 
 const (
-	_procPathCGroup     = "/proc/self/cgroup"
-	_procPathMountInfo  = "/proc/self/mountinfo"
-	_cgroupv2MountPoint = "/sys/fs/cgroup"
-
-	_cgroupV2CPUMaxDefaultPeriod = 100000
-	_cgroupV2CPUMaxQuotaMax      = "max"
-)
-
-const (
-	_cgroupv2CPUMaxQuotaIndex = iota
-	_cgroupv2CPUMaxPeriodIndex
+	_procPathCGroup    = "/proc/self/cgroup"
+	_procPathMountInfo = "/proc/self/mountinfo"
 )
 
 // CGroups is a map that associates each CGroup with its subsystem name.
@@ -141,70 +115,4 @@ func (cg CGroups) CPUQuota() (float64, bool, error) {
 	}
 
 	return float64(cfsQuotaUs) / float64(cfsPeriodUs), true, nil
-}
-
-// IsCGroupV2 returns true if the system supports and uses cgroup2.
-// It gets the required information for deciding from mountinfo file.
-func IsCGroupV2() (bool, error) {
-	return isCGroupV2(_procPathMountInfo)
-}
-
-func isCGroupV2(procPathMountInfo string) (bool, error) {
-	var isV2 bool
-	newMountPoint := func(mp *MountPoint) error {
-		if mp.FSType == _cgroupv2FSType && mp.MountPoint == _cgroupv2MountPoint {
-			isV2 = true
-		}
-		return nil
-	}
-	if err := parseMountInfo(procPathMountInfo, newMountPoint); err != nil {
-		return false, err
-	}
-	return isV2, nil
-}
-
-// CPUQuotaV2 returns the CPU quota applied with the CPU cgroup2 controller.
-// It is a result of reading cpu quota and period from cpu.max file.
-// It will return `cpu.max / cpu.period`. If cpu.max is set to max, it returns
-// (-1, false, nil)
-func CPUQuotaV2() (float64, bool, error) {
-	return cpuQuotaV2(_cgroupv2MountPoint, _cgroupv2CPUMax)
-}
-
-func cpuQuotaV2(cgroupv2MountPoint, cgroupv2CPUMax string) (float64, bool, error) {
-	cpuMaxParams, err := os.Open(path.Join(cgroupv2MountPoint, cgroupv2CPUMax))
-	if err != nil {
-		if os.IsNotExist(err) {
-			return -1, false, nil
-		}
-		return -1, false, err
-	}
-	scanner := bufio.NewScanner(cpuMaxParams)
-	if scanner.Scan() {
-		fields := strings.Fields(scanner.Text())
-		if len(fields) == 0 || len(fields) > 2 {
-			return -1, false, fmt.Errorf("invalid format")
-		}
-		if fields[_cgroupv2CPUMaxQuotaIndex] == _cgroupV2CPUMaxQuotaMax {
-			return -1, false, nil
-		}
-		max, err := strconv.Atoi(fields[_cgroupv2CPUMaxQuotaIndex])
-		if err != nil {
-			return -1, false, err
-		}
-		var period int
-		if len(fields) == 1 {
-			period = _cgroupV2CPUMaxDefaultPeriod
-		} else {
-			period, err = strconv.Atoi(fields[_cgroupv2CPUMaxPeriodIndex])
-			if err != nil {
-				return -1, false, err
-			}
-		}
-		return float64(max) / float64(period), true, nil
-	}
-	if err := scanner.Err(); err != nil {
-		return -1, false, err
-	}
-	return 0, false, io.ErrUnexpectedEOF
 }
